requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pandas>=2.0.0
sqlalchemy>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
gradio>=4.0.0

# API clients
biopython>=1.81
praw>=7.7.0

# Data processing
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
pydantic>=2.5.0

